backend/requirements.txt
fastapi
python-decouple
python-multipart
openai